requirements.txt
fastapi
pydantic
aiohttp
uvicorn
pyyaml